+++ src/LevelSize.hpp
@@ -2,5 +2,6 @@
 
 class LevelSize {
 public:
+    static bool get(std::string_view key);
     static std::string getSizeString(uint64_t size);
 };