+++ src/hooks/LevelCell.cpp
@@ -6,53 +6,44 @@ using namespace geode::prelude;
 
 class $modify(LSLevelCell, LevelCell) {
     static void onModify(ModifyBase<ModifyDerive<LSLevelCell, LevelCell>>& self) {
-        self.getHook("LevelCell::loadFromLevel").inspect([](Hook* hook) {
-            auto mod = Mod::get();
-            hook->setAutoEnable(mod->getSettingValue<bool>("show-size"));
+        if (auto it = self.m_hooks.find("LevelCell::loadFromLevel"); it != self.m_hooks.end()) {
+            auto hook = it->second.get();
+            hook->setAutoEnable(LevelSize::get("show-size"));
 
-            listenForSettingChangesV3<bool>("show-size", [hook](bool value) {
-                hook->toggle(value).inspectErr([](const std::string& err) {
-                    log::error("Failed to toggle LevelCell::loadFromLevel hook: {}", err);
-                });
-            }, mod);
-        }).inspectErr([](const std::string& err) {
-            log::error("Failed to get LevelCell::loadFromLevel hook: {}", err);
-        });
+            new EventListener([hook](std::shared_ptr<SettingV3> setting) {
+                if (auto res = hook->toggle(std::static_pointer_cast<BoolSettingV3>(std::move(setting))->getValue()); res.isErr()) {
+                    log::error("Failed to toggle LevelCell::loadFromLevel hook: {}", res.unwrapErr());
+                }
+            }, SettingChangedFilterV3(GEODE_MOD_ID, "show-size"));
+        }
     }
 
     void loadFromLevel(GJGameLevel* level) {
         LevelCell::loadFromLevel(level);
 
-        switch (level->m_levelType) {
-            case GJLevelType::Editor: {
-                auto sizeLabel = CCLabelBMFont::create(LevelSize::getSizeString(level->m_levelString.size()).c_str(), "goldFont.fnt");
-                sizeLabel->setPosition({ 350.0f, 3.0f });
-                sizeLabel->setScale(0.4f);
-                sizeLabel->setAnchorPoint({ 1.0f, 0.0f });
-                sizeLabel->setID("size-label"_spr);
-                m_mainLayer->addChild(sizeLabel);
-                break;
-            }
-            case GJLevelType::Saved: {
-                auto isDaily = level->m_dailyID.value() > 0;
-                auto whiteSize = isDaily || Mod::get()->getSettingValue<bool>("white-size");
-
-                auto sizeLabel = CCLabelBMFont::create(LevelSize::getSizeString(level->m_levelString.size()).c_str(), "chatFont.fnt");
-                sizeLabel->setPosition({ 346.0f + isDaily * 17.0f, 1.0f - isDaily * 6.0f });
-                sizeLabel->setScale(0.6f - m_compactView * 0.15f);
-                sizeLabel->setAnchorPoint({ 1.0f, 0.0f });
-                sizeLabel->setColor({
-                    (uint8_t)(51 * (whiteSize * 4 + 1)),
-                    (uint8_t)(51 * (whiteSize * 4 + 1)),
-                    (uint8_t)(51 * (whiteSize * 4 + 1))
-                });
-                sizeLabel->setOpacity(200 - whiteSize * 48);
-                sizeLabel->setID("size-label"_spr);
-                m_mainLayer->addChild(sizeLabel);
-                break;
-            }
-            default:
-                break;
+        auto levelType = level->m_levelType;
+        if (levelType == GJLevelType::Editor) {
+            if (!LevelSize::get("local-show-size")) return;
         }
+        else if (levelType == GJLevelType::Saved) {
+            if (!LevelSize::get("saved-show-size")) return;
+        }
+        else return;
+
+        auto sizeLabel = CCLabelBMFont::create(LevelSize::getSizeString(level->m_levelString.size()).c_str(), "chatFont.fnt");
+        if (levelType == GJLevelType::Saved) {
+            auto isDaily = level->m_dailyID.value() > 0;
+            auto whiteSize = isDaily || LevelSize::get("white-size");
+            sizeLabel->setPosition(isDaily ? CCPoint { 363.0f, -5.0f } : CCPoint { 346.0f, 1.0f });
+            sizeLabel->setColor(whiteSize ? ccColor3B { 255, 255, 255 } : ccColor3B { 51, 51, 51 });
+            sizeLabel->setOpacity(whiteSize ? 152 : 200);
+        }
+        else if (levelType == GJLevelType::Editor) {
+            sizeLabel->setPosition({ 346.0f, m_compactView ? 3.0f : 1.0f });
+        }
+        sizeLabel->setScale(m_compactView ? 0.45f : 0.6f);
+        sizeLabel->setAnchorPoint({ 1.0f, 0.0f });
+        sizeLabel->setID("size-label"_spr);
+        m_mainLayer->addChild(sizeLabel);
     }
 };